requirements.txt
flask==3.0.0
python-chess==1.10.0 